# contracts/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ShadowTrace",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/ingestion/source-records": {
      "post": {
        "tags": [
          "source"
        ],
        "summary": "Ingest Source Record",
        "operationId": "ingest_source_record_api_v1_ingestion_source_records_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestSourceRecordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestSourceRecordResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/source-records/{source_record_id}": {
      "get": {
        "tags": [
          "source"
        ],
        "summary": "Get Source Record",
        "operationId": "get_source_record_api_v1_source_records__source_record_id__get",
        "parameters": [
          {
            "name": "source_record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Record Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceRecordResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IngestSourceRecordRequest": {
        "properties": {
          "reference": {
            "$ref": "#/components/schemas/SourceReference"
          },
          "raw_payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Raw Payload"
          },
          "normalized": {
            "additionalProperties": true,
            "type": "object",
            "title": "Normalized"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "reference"
        ],
        "title": "IngestSourceRecordRequest"
      },
      "IngestSourceRecordResponse": {
        "properties": {
          "source_record_id": {
            "type": "string",
            "title": "Source Record Id"
          },
          "event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "accepted": {
            "type": "boolean",
            "title": "Accepted",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "source_record_id"
        ],
        "title": "IngestSourceRecordResponse"
      },
...
      "SourceDisposition": {
        "type": "string",
        "enum": [
          "pending",
          "processing",
          "contained",
          "completed",
          "suspended",
          "ignored",
          "unknown"
        ],
        "title": "SourceDisposition",
        "description": "Normalized external disposition label; source_status_raw keeps the original."
      },
      "SourceObjectKind": {
        "type": "string",
        "enum": [
          "incident",
          "alert",
          "asset",
          "log",
          "connector"
        ],
        "title": "SourceObjectKind",
        "description": "Canonical source object kind used for the internal discriminated union."
      },
...
      "SourceRecordResponse": {
        "properties": {
          "source_record_id": {
            "type": "string",
            "title": "Source Record Id"
          },
          "reference": {
            "$ref": "#/components/schemas/SourceReference"
          },
          "normalized": {
            "additionalProperties": true,
            "type": "object",
            "title": "Normalized"
          },
          "current_source_disposition": {
            "$ref": "#/components/schemas/SourceDisposition",
            "default": "unknown"
          },
          "source_sync_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Sync State"
          }
        },
        "type": "object",
        "required": [
          "source_record_id",
          "reference"
        ],
        "title": "SourceRecordResponse"
      },
      "SourceReference": {
        "properties": {
          "source_kind": {
            "$ref": "#/components/schemas/SourceObjectKind"
          },
          "source_product": {
            "type": "string",
            "title": "Source Product"
          },
          "source_tenant_id": {
            "type": "string",
            "title": "Source Tenant Id"
          },
          "connector_id": {
            "type": "string",
            "title": "Connector Id"
          },
          "source_object_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Object Type"
          },
          "source_object_id": {
            "type": "string",
            "title": "Source Object Id"
          },
          "parent_source_object_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Source Object Id"
          },
          "source_status_raw": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Status Raw"
          },
          "source_disposition": {
            "$ref": "#/components/schemas/SourceDisposition",
            "default": "unknown"
          },
          "source_concurrency_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Concurrency Token"
          },
          "source_updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Updated At"
          },
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "1"
          },
          "ingested_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ingested At"
          },
          "raw_payload_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Payload Hash"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source_kind",
          "source_product",
          "source_tenant_id",
          "connector_id",
          "source_object_id"
        ],
        "title": "SourceReference",
        "description": "Immutable snapshot reference to an external source object (intro §4.3).\n\nIdentity five-tuple is\n``(source_product, source_tenant_id, connector_id, source_kind, source_object_id)``.\nThe adapter-native ``source_object_type`` and opaque ``source_concurrency_token``\ndo NOT participate in identity."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}